requirements.txt
bitstring==4.0.1
cryptography==38.0.4
pycryptodome==3.16.0
requests==2.22.0




